ENH: Make the integer conversions in allocations, copies and reads explicit
152 of the -Wsign-conversion findings are an int or int64_t count
reaching a size_t parameter of malloc(), calloc(), realloc(), memcpy(),
memset(), strncpy(), znzread(), znzwrite() or fread().

Where a count is multiplied by a sizeof, the cast goes on the count
rather than around the product, so the multiply happens at 64-bit width:
(int)count * sizeof(T) can overflow before it is widened.

The counts cannot be negative. nifti_update_dims_from_array() clamps
every dim[i] to at least 1, nifti_datatype_sizes() leaves nbyper at 0
for an unknown datatype, and nifti_image_load() requires nbyper > 0 and
nvox > 0.

+++ cifti/afni_xml.c
@@ -196,7 +196,7 @@ afni_xml_list axml_read_file(const char * fname, int read_data)
    {
       if( reset_xml_buf(xd, &buf, &bsize) ) break;
 
-      blen = fread(buf, 1, bsize, fp);
+      blen = fread(buf, 1, (size_t)bsize, fp);
 
       /* check for early termination */
       bshort = loc_strnlen(buf, blen);
@@ -447,8 +447,8 @@ int axml_add_attrs(afni_xml_t * ax, const char ** attr)
       return 0;
    }
 
-   ax->attrs.name = (char **)malloc(natr*sizeof(char *));
-   ax->attrs.value = (char **)malloc(natr*sizeof(char *));
+   ax->attrs.name = (char **)malloc((size_t)natr * sizeof(char *));
+   ax->attrs.value = (char **)malloc((size_t)natr * sizeof(char *));
 
    /* failure? */
    if( ! ax->attrs.name || ! ax->attrs.value ) {
@@ -666,7 +666,7 @@ static int reset_xml_buf(afni_xml_control * xd, char ** buf, int * bsize)
         fprintf(stderr,"++ update buf, %d to %d bytes\n",*bsize,xd->buf_size);
 
     *bsize = xd->buf_size;
-    *buf = (char *)safe_realloc(*buf, (*bsize+1) * sizeof(char));
+    *buf = (char *)safe_realloc(*buf, (size_t)((*bsize+1) * sizeof(char)));
     if( ! *buf ) {
         fprintf(stderr,"** failed to alloc %d bytes of xml buf!\n", *bsize);
         *bsize = 0;
@@ -826,7 +826,7 @@ static int add_to_xroot_list(afni_xml_control * xd, afni_xml_t * newp)
 
    xd->xroot->len++;
    xd->xroot->xlist = (afni_xml_t **)safe_realloc(xd->xroot->xlist,
-                                     xd->xroot->len * sizeof(afni_xml_t *));
+                                     (size_t)xd->xroot->len * sizeof(afni_xml_t *));
    if( ! xd->xroot->xlist ) {
       fprintf(stderr,"** failed to alloc %d AXMLT pointers\n", xd->xroot->len);
       return 1;
@@ -844,7 +844,7 @@ static int add_to_xchild_list(afni_xml_t * parent, afni_xml_t * child)
 
    parent->nchild++;
    parent->xchild = (afni_xml_t **)safe_realloc(parent->xchild,
-                                   parent->nchild * sizeof(afni_xml_t *));
+                                   (size_t)parent->nchild * sizeof(afni_xml_t *));
    if( ! parent->xchild ) {
       fprintf(stderr,"** failed to alloc %d AXML pointers\n", parent->nchild);
       return 1;
@@ -906,7 +906,7 @@ static const char * strip_whitespace(const char * str, int slen)
 
    /* make sure we have local space */
    if( len > blen ) { /* allocate a bigger buffer */
-      buf = (char *)safe_realloc(buf, (len+1) * sizeof(char));
+      buf = (char *)safe_realloc(buf, (size_t)((len+1) * sizeof(char)));
       if( !buf ) {
          fprintf(stderr,"** failed to alloc wspace buf of len %d\n", len+1);
          return str;
@@ -919,7 +919,7 @@ static const char * strip_whitespace(const char * str, int slen)
 
    if( ifirst == len ) *buf = '\0';
    else {
-      strncpy(buf, str+ifirst, len-ifirst-ilast);
+      strncpy(buf, str+ifirst, (size_t)(len-ifirst-ilast));
       buf[len-ifirst-ilast] = '\0';
    }
 
@@ -971,14 +971,14 @@ static int append_to_string(char ** ostr, int * olen,
 
    newlen = *olen + ilen;
 
-   *ostr = (char *)safe_realloc(*ostr, newlen * sizeof(char));
+   *ostr = (char *)safe_realloc(*ostr, (size_t)newlen * sizeof(char));
    if( !*ostr ) {
       fprintf(stderr,"** AX.A2S: failed to alloc %d chars\n", newlen);
       return 1;
    }
 
    /* copy, starting at old nul char (if any), and terminate */
-   strncpy((*ostr)+*olen-1, istr, ilen);
+   strncpy((*ostr)+*olen-1, istr, (size_t)ilen);
    (*ostr)[newlen-1] = '\0';
    *olen = newlen;
